+++ src/miner.cpp
@@ -15,19 +15,25 @@
 #include "program_util.hpp"
 #include "setup.hpp"
 
+const std::string Miner::ANONYMOUS("anonymous");
+
 Miner::Miner(const Settings &settings) : settings(settings) {}
 
-void Miner::reload() {
+void Miner::reload(bool load_generators) {
   manager.reset(new OeisManager(settings));
   manager->load();
   manager->getFinder();  // initializes matchers
-  multi_generator.reset(
-      new MultiGenerator(settings, manager->getStats(), true));
+  if (load_generators) {
+    multi_generator.reset(
+        new MultiGenerator(settings, manager->getStats(), true));
+  } else {
+    multi_generator.reset();
+  }
 }
 
-void Miner::mine(const std::vector<std::string> &initial_progs) {
+void Miner::mine() {
   if (!manager) {
-    reload();
+    reload(true);
   }
   ApiClient api_client;
   Mutator mutator;
@@ -38,14 +44,6 @@ void Miner::mine(const std::vector<std::string> &initial_progs) {
   AdaptiveScheduler metrics_scheduler(Metrics::get().publish_interval);
   AdaptiveScheduler api_scheduler(600);       // 10 minutes (magic number)
   AdaptiveScheduler reload_scheduler(43200);  // 12 hours (magic number)
-  static const std::string anonymous("anonymous");
-
-  // load initial programs
-  for (auto &path : initial_progs) {
-    program = parser.parse(path);
-    progs.push(program);
-    mutator.mutateConstants(program, 1000, progs);
-  }
 
   // get mining mode
   const auto mode = Setup::getMiningMode();
@@ -75,30 +73,19 @@ void Miner::mine(const std::vector<std::string> &initial_progs) {
       progs.push(generator->generateProgram());
     }
 
-    // get next program and match sequences
+    // get next program
     program = progs.top();
+    progs.pop();
     // Log::get().info( "Matching program with " + std::to_string(
     // program.ops.size() ) + " operations" ); ProgramUtil::print( program,
     // std::cout );
-    progs.pop();
+
+    // match sequences
     auto seq_programs = manager->getFinder().findSequence(
         program, norm_seq, manager->getSequences());
     for (auto s : seq_programs) {
       program = s.second;
-
-      // update "submitted by" comment
-      submitted_by = ProgramUtil::getSubmittedBy(program);
-      if (submitted_by.empty()) {
-        submitted_by = Setup::getSubmittedBy();
-        if (!submitted_by.empty()) {
-          Operation nop(Operation::Type::NOP);
-          nop.comment = ProgramUtil::SUBMITTED_BY_PREFIX + " " + submitted_by;
-          program.ops.push_back(nop);
-        }
-      } else if (submitted_by == anonymous) {
-        ProgramUtil::removeOps(program, Operation::Type::NOP);
-      }
-
+      setSubmittedBy(program);
       auto r = manager->updateProgram(s.first, program);
       if (r.first) {
         // update stats and increase priority of successful generator
@@ -127,13 +114,7 @@ void Miner::mine(const std::vector<std::string> &initial_progs) {
         if (program.ops.empty()) {
           break;
         }
-        // update "submitted by" comment
-        submitted_by = ProgramUtil::getSubmittedBy(program);
-        if (submitted_by.empty()) {
-          Operation nop(Operation::Type::NOP);
-          nop.comment = ProgramUtil::SUBMITTED_BY_PREFIX + " " + anonymous;
-          program.ops.push_back(nop);
-        }
+        ensureSubmittedBy(program);
         progs.push(program);
       }
     }
@@ -146,16 +127,12 @@ void Miner::mine(const std::vector<std::string> &initial_progs) {
       for (size_t i = 0; i < multi_generator->generators.size(); i++) {
         auto gen = multi_generator->generators[i].get();
         auto labels = gen->metric_labels;
-
         labels["kind"] = "generated";
         entries.push_back({"programs", labels, (double)gen->stats.generated});
-
         labels["kind"] = "new";
         entries.push_back({"programs", labels, (double)gen->stats.fresh});
-
         labels["kind"] = "updated";
         entries.push_back({"programs", labels, (double)gen->stats.updated});
-
         total_generated += gen->stats.generated;
         gen->stats = Generator::GStats();
       }
@@ -168,8 +145,81 @@ void Miner::mine(const std::vector<std::string> &initial_progs) {
     // regular task: reload oeis manager and generators
     if (reload_scheduler.isTargetReached()) {
       reload_scheduler.reset();
-      reload();
+      reload(true);
       generator = multi_generator->getGenerator();
     }
+  }
+}
+
+void Miner::submit(const std::string &id, const std::string &path) {
+  Parser parser;
+  auto program = parser.parse(path);
+  if (!manager) {
+    reload(false);
+  }
+  OeisSequence seq(id);
+  Settings settings(this->settings);
+  settings.print_as_b_file = false;
+  Log::get().info("Validating program for " + seq.id_str());
+  Evaluator evaluator(settings);
+  auto terms = seq.getTerms(100000);  // magic number
+  auto result =
+      evaluator.check(program, terms, OeisSequence::DEFAULT_SEQ_LENGTH, seq.id);
+  if (result.first == status_t::ERROR) {
+    Log::get().error("Validation failed", false);
+    settings.print_as_b_file = true;
+    Evaluator evaluator2(settings);
+    evaluator2.check(program, terms, OeisSequence::DEFAULT_SEQ_LENGTH, seq.id);
+    return;  // error
+  }
+  Log::get().info("Validation successful");
+  // match sequences
+  ApiClient api_client;
+  Sequence norm_seq;
+  const auto mode = Setup::getMiningMode();
+  auto seq_programs = manager->getFinder().findSequence(
+      program, norm_seq, manager->getSequences());
+  size_t matches = 0;
+  for (auto s : seq_programs) {
+    program = s.second;
+    setSubmittedBy(program);
+    auto r = manager->updateProgram(s.first, program);
+    if (r.first) {
+      // in client mode: submit the program to the API server
+      if (mode == MINING_MODE_CLIENT) {
+        api_client.postProgram(program);
+      }
+      matches++;
+    }
+  }
+  if (matches > 0) {
+    Log::get().info("Submitted programs for " + std::to_string(matches) +
+                    " matched sequences");
+  } else {
+    Log::get().info(
+        "Skipped submission because there exists already a (faster) program");
+  }
+}
+
+void Miner::ensureSubmittedBy(Program &program) {
+  auto submitted_by = ProgramUtil::getSubmittedBy(program);
+  if (submitted_by.empty()) {
+    Operation nop(Operation::Type::NOP);
+    nop.comment = ProgramUtil::SUBMITTED_BY_PREFIX + " " + ANONYMOUS;
+    program.ops.push_back(nop);
+  }
+}
+
+void Miner::setSubmittedBy(Program &program) {
+  auto submitted_by = ProgramUtil::getSubmittedBy(program);
+  if (submitted_by.empty()) {
+    submitted_by = Setup::getSubmittedBy();
+    if (!submitted_by.empty()) {
+      Operation nop(Operation::Type::NOP);
+      nop.comment = ProgramUtil::SUBMITTED_BY_PREFIX + " " + submitted_by;
+      program.ops.push_back(nop);
+    }
+  } else if (submitted_by == ANONYMOUS) {
+    ProgramUtil::removeOps(program, Operation::Type::NOP);
   }
 }