+++ src/oeis_maintenance.cpp
@@ -116,29 +116,6 @@ void OeisMaintenance::generateLists() {
   no_loda_file << no_loda.str();
   no_loda_file.close();
 
-  // publish metrics
-  auto &stats = manager.getStats();
-  std::vector<Metrics::Entry> entries;
-  std::map<std::string, std::string> labels;
-
-  labels["kind"] = "total";
-  entries.push_back({"programs", labels, (double)num_processed});
-  entries.push_back({"sequences", labels, (double)manager.getTotalCount()});
-
-  labels["kind"] = "used";
-  entries.push_back({"sequences", labels, (double)stats.num_sequences});
-
-  labels.clear();
-  for (size_t i = 0; i < stats.num_ops_per_type.size(); i++) {
-    if (stats.num_ops_per_type[i] > 0) {
-      labels["type"] =
-          Operation::Metadata::get(static_cast<Operation::Type>(i)).name;
-      entries.push_back(
-          {"operation_types", labels, (double)stats.num_ops_per_type[i]});
-    }
-  }
-  Metrics::get().write(entries);
-
   Log::get().info("Finished generation of lists for " +
                   std::to_string(num_processed) + " programs");
 }